requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
billiard==4.1.0
black==23.7.0
celery==5.3.1
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
Django==4.2.4
django-celery-beat==2.5.0
django-debug-toolbar==4.1.0
django-timezone-field==6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.4
flake8==6.1.0
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kombu==5.3.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
Pillow==10.0.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
pycodestyle==2.11.0
pyflakes==3.1.0
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
redis==5.0.0
referencing==0.30.2
rpds-py==0.9.2
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
vine==5.0.0
wcwidth==0.2.6
